+++ src/extension.ts
@@ -2,13 +2,21 @@ import * as vscode from 'vscode';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as os from 'os';
+import * as chokidar from 'chokidar';
 import * as Color from 'color';
 import template from './template';
 
+<<<<<<< HEAD
 const walCachePath = path.join(os.homedir(), '/.cache/wal');
 const walColorsPath = path.join(walCachePath, '/colors');
 const walColorsJsonPath = path.join(walCachePath, '/colors.json');
 let autoUpdateWatcher: fs.FSWatcher | null = null;
+=======
+const walCachePath = path.join(os.homedir(), '.cache', 'wal');
+const walColorsPath = path.join(walCachePath, 'colors');
+const walColorsJsonPath = path.join(walCachePath, 'colors.json');
+let autoUpdateWatcher: chokidar.FSWatcher | null = null;
+>>>>>>> da5c232de8d0859eb12f1a495c36c5267a992126
 
 
 export function activate(context: vscode.ExtensionContext) {
@@ -62,7 +70,7 @@ export function deactivate() {
  */
 function generateColorThemes() {
 	// Import colors from pywal cache
-	let colors: Color[];
+	let colors: Color[] | undefined;
 	try {
 		colors = fs.readFileSync(walColorsPath)
 										 .toString()
@@ -77,36 +85,56 @@ function generateColorThemes() {
 				}
 			};
 
+<<<<<<< HEAD
 			const colorsJson: WalJson = JSON.parse(
 					fs.readFileSync(walColorsJsonPath)
 						.toString()
 			);
+=======
+			let colorsJson: WalJson;
+			const colorsRaw = fs.readFileSync(walColorsJsonPath).toString();
+
+			try {
+				colorsJson = JSON.parse(colorsRaw);
+			} catch {
+				// The wallpaper path on Windows can cause JSON.parse errors since the
+				// path isn't properly escaped.
+				colorsJson = JSON.parse(colorsRaw
+					.split('\n')
+					.filter((line) => !line.includes('wallpaper'))
+					.join('\n'));
+			}
+>>>>>>> da5c232de8d0859eb12f1a495c36c5267a992126
 
 			colors[0] = Color(colorsJson?.special?.background);
 			colors[7] = Color(colorsJson?.special?.foreground);
 		}
 	} catch(error) {
-		vscode.window.showErrorMessage('Couldn\'t load colors from pywal cache, be sure to run pywal before updating.');
-		return;
+		// Not a complete failure if we have colors from the wal colors file, but failed to load from the colors.json
+		if (colors === undefined || colors.length === 0) {
+			vscode.window.showErrorMessage('Couldn\'t load colors from pywal cache, be sure to run pywal before updating.');
+			return;
+		}
+
+		vscode.window.showWarningMessage('Couldn\'t load all colors from pywal cache');
 	}
 		
 	// Generate the normal theme
 	const colorTheme = template(colors, false);
-	fs.writeFileSync(path.join(__dirname,'../themes/wal.json'), JSON.stringify(colorTheme, null, 4));
+	fs.writeFileSync(path.join(__dirname,'..', 'themes', 'wal.json'), JSON.stringify(colorTheme, null, 4));
 	
 	// Generate the bordered theme
 	const colorThemeBordered = template(colors, true);
-	fs.writeFileSync(path.join(__dirname,'../themes/wal-bordered.json'), JSON.stringify(colorThemeBordered, null, 4));
+	fs.writeFileSync(path.join(__dirname,'..', 'themes', 'wal-bordered.json'), JSON.stringify(colorThemeBordered, null, 4));
 }
 
 /**
  * Automatically updates the theme when the color palette changes
  * @returns The watcher for the color palette
  */
-function autoUpdate(): fs.FSWatcher {
-	let fsWait = false;
-
+function autoUpdate(): chokidar.FSWatcher {
 	// Watch for changes in the color palette of wal
+<<<<<<< HEAD
 	return fs.watch(walCachePath, (event, filename) => {
 		if (filename) {
 			// Delay after a change is found
@@ -122,4 +150,9 @@ function autoUpdate(): fs.FSWatcher {
 			generateColorThemes();
 		}
 	});
+=======
+	return chokidar
+		.watch(walCachePath)
+		.on('change', generateColorThemes);
+>>>>>>> da5c232de8d0859eb12f1a495c36c5267a992126
 }